teamLogをsetする箇所を共通化
